# AtomicCell: Use atomic-maybe-uninit