--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -62,6 +62,10 @@ jobs:
           - rust: nightly
             os: ubuntu-latest
             target: armv5te-unknown-linux-gnueabi
+          # Test target without stable inline asm support.
+          - rust: stable
+            os: ubuntu-latest
+            target: sparc64-unknown-linux-gnu
     runs-on: ${{ matrix.os }}
     timeout-minutes: 60
     steps: