This fixes UB when lock-free and the value has uninitialized bytes, by using [atomic-maybe-uninit](https://github.com/taiki-e/atomic-maybe-uninit) crate.
Fixes #315
Fixes #748

Also, this makes 128-bit atomics lock-free on some platforms.
Closes https://github.com/crossbeam-rs/rfcs/issues/13

Currently, this patch always uses fallback (seqlock) on environments that do not support inline assembly, such as Miri.
It is possible to use atomic-maybe-uninit crate to fix UB of concurrent access with volatile read/write in seqlock/deque (#744, #859), but that is not included in this PR and is left to subsequent PRs.

---

cc @RalfJung