@taiki-e 2023-08-14
Filed https://github.com/rust-lang/rust/pull/114790 to allow MaybeUninit in input and output of inline assembly.