@taiki-e 2024-12-09
Since this PR was opened we have made many improvements to both atomic-maybe-uninit and the compiler. Resolving performance issue (https://github.com/rust-lang/rust/pull/114790), improving and stabilizing inline assembly support on tier 2 targets (https://github.com/rust-lang/rust/pull/131258, https://github.com/rust-lang/rust/pull/131781, etc.), fixing or avoiding rustc/LLVM bugs, improving code quality (https://github.com/taiki-e/atomic-maybe-uninit/pull/30, etc.), improving platform support (atomic-maybe-uninit currently supports all CPU architectures supported by Rust other than C-SKY), etc.

I think we can now finally merge this.